+++ vectoria_lib/preprocessing/document_data.py
